Spacial ICA for fmri data

Uses fastICA to perform spatial ICA on fMRI data.

fmri.sICA(data, mask=NULL, ncomp=20,
  alg.typ=c("parallel","deflation"), fun=c("logcosh","exp"),
  alpha=1, detrend=TRUE, degree=2, nuisance= NULL, ssmooth=TRUE,
  tsmooth=TRUE, bwt=4, bws=8, unit=c("FWHM","SD"))

Arguments

data
fMRI dataset of class ''fmridata''

mask
Brain mask, if NULL then data$mask is used.

ncomp
Number of ICA components to compute.

alg.typ
Alg. to be used in fastICA.

fun
Test functions to be used in fastICA.

alpha
Scale parameter in test functions, see fastICA.

detrend
Trend removal (polynomial)

degree
degree of polynomial trend

nuisance
Matrix of additional nuisance parameters to regress against.

ssmooth
Should spatial smoothing be used for variance reduction

tsmooth
Should temporal smoothing be be applied

bws
Bandwidth for spatial Gaussian kernel

bwt
Bandwidth for temporal Gaussian kernel

unit
Unit of bandwidth, either standard deviation (SD) of Full Width Half Maximum (FWHM).

Details

If specified polynomial trends and effects due to nuisance parameters, e.g., motion parameters, are removed. If smooth==TRUE the resulting residual series is spatially smoothed using a Gaussian kernel with specified bandwidth. ICA components are the estimated using fastICA based on data within brain mask. The components of the result are related as XKW=scomp[mask,] and X=scomp[mask,]*A.

Value

object of class ''fmriICA'' list with components

scomp

4D array with ICA component images. Last index varies over components.

X

pre-processed data matrix

K

pre-processed data matrix

W

estimated un-mixing matrix

A

estimated mixing matrix

mask

Brain mask

pixdim

voxelsize

TR

Repetition Time (TR)

Author

Joerg Polzehl polzehl@wias-berlin.de

See also