Searchlight signal detection

Detection of activated regions using searchlights.

fmri.searchlight(spm, alpha = 0.05, radius, minimum.signal = 0,
   kind = c("abs", "squared"))

Arguments

spm
fmrispm object

alpha
multiple test (over volume) adjusted significance level.

radius
radius of searchlight. Value needs to be larger or eaqual than 1.

minimum.signal
allows to specify a (positive) minimum value for detected signals. If minimum.signal >0 the thresholds are to conservative, this case needs further improvements.

kind
Kind of statistics used for aggregation over search light region. "abs" specifies averaging of absolute voxelwise t-statistics while "squared" corresponds to averaging of squares of these statistics.

Details

The function computes mean statistics (depending on kind) over a searchlight region of radius radius. Approximate voxelwise p-values are determined with respect an empirical (simulated) distribution of the searchlight statistics under the null hypothesis a central t-distributed spm. Thresholding used FDR with rate alpha.

Value

Object with class attributes "fmripvalue" and "fmridata"

pvalue

voxelwise p-value if exceeding FDR-critical value, 1 otherwise.

weights

voxelsize ratio

dim

data dimension

hrf

expected BOLD response for contrast (single stimulus only)

References

Kriegeskorte, N.; Goebel, R. & Bandettini, P. (2006) Information-based functional brain mapping, PNAS 103:3863-3868.

Author

Joerg Polzehl polzehl@wias-berlin.de

See also

Examples

  if (FALSE) fmri.searchlight(fmrispmobj)