Spatial group ICA for fmri

Combine ICA results from multiple runs or multiple subjects in group fMRI studies

fmri.sgroupICA(icaobjlist, thresh = 0.75, minsize=2)

Arguments

icaobjlist
List of results obtained by function fmri.sICA for a series of fmri data sets (multiple runs or multiple subjects).

thresh
threshold for cluster aggregation. Needs to be in (0,1).

minsize
Minimal size of cluster to consider in IC aggregation. Needs to be larger than 1.

Details

The fMRI time series need to be preprocessed and registered before thr ICA decomposition is performed.

The function employes a hierarchical clustering algorithm (complete linkage) on the combined set of spatial independent components obtained from the individual time series. A distance matrix is obtained from correlations of the independent component images. Aggregation of two components from the same fmri series is prevented in the algorithm.

Value

An object of class ''fmrigroupICA'' with components

icacomp

Mean IC's over cluster members for cluster of size larger or equal minsize

size

Size of selected clusters

cl

Number of selected clusters

cluster

Cluster membership corresponding to thresh.

height

Distance value at which the cluster was createt. Elements correspond to elements of cluster.

hdm

Object returned by function hclust.

References

F. Esposito et al (2005) Independent component analysis of fMRI group studies by self-organizing clustering, Neuroimage, pp. 193-205.

Author

Joerg Polzehl polzehl@wias-berlin.de

See also