Create fmridata object from niftiImage

Transforms a niftiImage (created by readNifti from package RNiftyReg) into an object with class fmridata

niftiImage2fmri(niftiobj, level = 0.75, mask=NULL, setmask = TRUE, indx = NULL,
   indy = NULL, indz = NULL, avoidnegs = FALSE)

Arguments

niftiobj
an object of class niftiImage

level
quantile used in mask definition

mask
array or nifti-object containing the mask. If set this replaces the mask defined by argument level.

setmask
if TRUE create a brain mask

indx
index vector for subcube definition

indy
index vector for subcube definition

indz
index vector for subcube definition

avoidnegs
if TRUE change the mean to avoid negative image intensities

Details

This function can be used in connection with readNifti from package RNiftyReg to read large fMRI series from nifti files. The resulting fmridata-object stores the image data as 2 byte integer in raw format, in contrast for the 4 byte real used with other functions.

Value

an object of class fmridata

Author

Joerg Polzehl polzehl@wias-berlin.de

See also