Convert Between fmridata and oro.nifti Objects

NIfTI data can be converted between fmridata S3 objects (from the fmri package) and nifti S4 objects.

oro2fmri(from, value = NULL, level = 0.75, mask=NULL, setmask = TRUE)
fmri2oro(from, value = NULL, verbose = FALSE, reorient = FALSE,
         call = NULL)

Arguments

from
is the object to be converted.

value
NULL

level
is the quantile level defining the mask.

mask
array or nifti-object containing the mask. If set this replaces the mask defined by argument level.

setmask
is a logical variable (default = TRUE), whether to define a suitable mask based on level.

verbose
is a logical variable (default = FALSE) that allows text-based feedback during execution of the function.

reorient
is a logical variable (default = TRUE) that enforces Qform/Sform transformations.

call
keeps track of the current function call for use in the NIfTI extension.

Details

These functions enhance the capabilities of fmri by allowing the exchange of data objects between nifti and fmridata classes.

Value

The function oro2fmri produces an S3 object of class fmridata. The function fmri2oro produces an S4 object of class nifti.

See also

Author

Brandon Whitcher bwhitcher@gmail.com