I/O functions

Visualize fMRI data and (intermediate) results.

# S3 method for fmridata
plot(x, anatomic = NULL, maxpvalue = 0.05,
              spm = TRUE, pos = c(-1, -1, -1), type = "slice",
              slice =  1, view = "axial" ,zlim.u =
              NULL, zlim.o = NULL,col.o = heat.colors(256), col.u =
              grey(0:255/255), cutOff = c(0, 1), ...)
# S3 method for fmrisegment
plot(x, anatomic = NULL,
              slice =  1, view = c( "axial", "coronal", "sagittal") ,zlim.u =
              NULL, zlim.o = NULL,col.o = c( rainbow( 64, start = 2/6, end = 4/6),
              rainbow( 64, start = 0, end = 1/6)),
              col.u = grey(0:127/127), verbose = FALSE, ...)

Arguments

x
object of class "fmrisegment", "fmrispm" or "fmridata"

anatomic
overlay of same dimension as the functional data, or fmridata object (if of x is fmripvalue object)

maxpvalue
maximum p-value for thresholding

spm
logical. if class is "fmrispm" decide whether to plot the t-statistics for the estimated effect (spm=TRUE) or the estimated effect itself (spm=FALSE).

pos
voxel to be marked on output

type
string. "slice" for slicewise view and "3d" for 3d view.

slice
number of slice in x, if anatomic is of "fmridata" class

view
"axial", "coronal", or "sagittal", if anatomic is of "fmridata" class

zlim.u
full range for anatomical underlay used for color scale, if anatomic is of "fmridata" class

zlim.o
full range for functional overlay used for color scale, if anatomic is of "fmridata" class

col.u
color scale for anatomical underlay, if anatomic is of "fmridata" class, default grey(0:255/255)

col.o
color scale for functional overlay, if anatomic is of "fmridata" class, default heat.colors(256)

cutOff
not yet documented

verbose
tell something on the progress?

...
additional arguments for plot

Details

Provides a sliceswise view of "fmridata" objects with anatomic overlay (if appropriate, that is for class "fmripvalue"). For objects of class "fmrispm" it plots the t-statistics for the estimated effects if spm is TRUE, or the estimated effect otherwise. For objects of class "fmridata" only a plot of the data slices itself is produced. If device is specified as "png", "jpeg", "ppm" output is done to a file. A grey/color scale is provided in the remaining space.

For objects of class "fmrisegment" the smoothed signal size is shown in the activation segments (two-sided test!).

If type is "3d" a 3 dimensional interactive view opens. Sliders to move in the data cube are given ("x", "y", "z", and "t" if class is "fmridata" only). Time series are shown if available. For objects of class "fmrispm" a slider is created to remove information for voxels with smaller signals than a cut-off value from the plot. Use pvalues for statistical evaluation. If spm is FALSE the estimated BOLD response together with a confidence interval corresponding to maxpvalue is drawn. For objects of class "fmripvalue" the pvalues with overlay are shown.

Value

If 'type' is "3d" the Tk-object is returned. (Remove the diplay with tkdestroy(object))

References

Polzehl, J. and Tabelow, K. (2007) fmri: A Package for Analyzing fmri Data, R News, 7:13-17 .

Author

Karsten Tabelow tabelow@wias-berlin.de

Note

3 dimensional plotting requires the tkrplot package.

See also

Examples

if (FALSE) plot(pvalue)