I/O function

Read HEAD/BRIK file.

read.AFNI(filename,vol=NULL,level=0.75,mask=NULL,setmask=TRUE)

Arguments

filename
name of the file (without extension)

vol
vector of volumes of the dataset to be read

level
Quantile level defining the mask

mask
array or nifti-object containing the mask. If set this replaces the mask defined by argument level.

setmask
Logical (default TRUE), whether to define a suitable mask based on level

Details

The function reads a HEAD/BRIK file. If vol is given (defaults to NULL), only volumes in this vector are read, in order to save memory.

Value

Object of class "fmridata" with the following list entries:

ttt

raw vector (numeric size 4) containing the four dimensional data cube (the first three dimensions are voxel dimensions, the fourth dimension denotes the time).

header

header information list

format

data source. string "HEAD/BRIK"

delta

voxel size in mm

origin

position of the datacube origin

orient

data orientation code. see AFNI documentation

dim

dimension of the datacube

weights

weights vector coding the relative voxel sizes in x, y, z-direction.

mask

head mask

Author

Karsten Tabelow tabelow@wias-berlin.de

References

R. W. Cox (1996). AFNI: Software for analysis and visualization of functional magnetic resonance neuroimages. Computers and Biomed. Res. 29:162-173.

Polzehl, J. and Tabelow, K. (2007) fmri: A Package for Analyzing fmri Data, R News, 7:13-17 .

See also

Examples

if (FALSE) afni <- read.AFNI("afnifile")