I/O functions

Write BRIK/HEAD files.

write.AFNI(filename, ttt, label = NULL, note = NULL, origin = NULL, 
               delta = NULL, idcode = NULL, header = NULL, taxis = FALSE)

Arguments

filename
name of the file

ttt
datacube

label
labels (BRICK\_LABS), depreciated - see header

note
notes on data (HISTORY\_NOTE), depreciated - see header

origin
origin of datacube (ORIGIN), depreciated - see header

delta
voxel dimensions (DELTA), depreciated - see header

idcode
idcode of data (IDCODE\_STRING), depreciated - see header

header
This is a list of header information such as DATASET\_RANK to be written to the .HEAD file. Arguments label, ... are depreciated and to be substituted by a corresponding list entry. For backward compatibility the use of the old arguments is still supported and should give the same results. This will be removed in some future release! Since AFNI does not read any dataset with a header choose carefully what is written. There are some basic tests in this function, but this may not be sufficient.

taxis
logical (defaults to FALSE. Are the sub-bricks time series? This results in writing TAXIS attributes to the header file.

Details

Write out BRIK/HEAD files as required by AFNI. Most arguments correspond to entries in the HEAD file, but use is depreciated. Use header and taxis instead!

Value

Nothing is returned.

References

Polzehl, J. and Tabelow, K. (2007) fmri: A Package for Analyzing fmri Data, R News, 7:13-17 .

Author

Karsten Tabelow tabelow@wias-berlin.de

See also

Examples

 if (FALSE) write.AFNI(tempfile(), array(as.integer(65526*runif(10*10*10*20)),
     c(10,10,10,20)), c("signal"), note="random data",
     origin=c(0,0,0), delta=c(4,4,5), idcode="unique ID")
 write.AFNI(tempfile(), array(as.integer(65526*runif(10*10*10*20)),
     c(10,10,10,20)), header=list(HISTORY_NOTE="random data",
     ORIGIN=c(0,0,0), DELTA=c(4,4,5), IDCODE_STRING="unique ID"),taxis=FALSE)
#> Warning: TYPESTRING not given, setting to 3DIM_HEAD_FUNC for backward compatibility
#> Warning: SCENE_DATA not given, setting to c(0,11,1,-999,-999,-999,-999,-999) for backward compatibility
#> Warning: ORIENT_SPECIFIC not given, setting to c(0,3,4) for backward compatibility
#> Warning: no BRICK_TYPES given, assuming short